Don't show table if no study guides in admin

+++ admin/manage-study-guides.php
@@ -15,26 +15,28 @@
 
 <div id="manage-study-guides">
     <a class="btn waves-effect waves-light" href="upload-study-guide.php">Upload Study Guide</a>
-    <table class="striped">
-        <thead>
-            <tr>
-                <th>Display Name</th>
-                <th>Rename</th>
-                <th>Delete</th>
-            </tr>
-        </thead>
-        <tbody>
-            <?php foreach ($files as $file) { ?>
+    <?php if (count($files) !== 0) { ?>
+        <table class="striped">
+            <thead>
                 <tr>
-                    <td><a target="_blank" class="btn-flat blue-text waves-effect waves-blue no-uppercase" href="<?= $basePath . '/' .$file['FileName'] ?>"><?= $file['DisplayName'] ?></a></td> 
-                    <td><a class="btn waves-effect" 
-                            href="rename-study-guide.php?id=<?= $file['StudyGuideID'] ?>">Rename</a></td> 
-                    <td><a class="btn red white-text waves-effect waves-light" 
-                            href="delete-study-guide.php?id=<?= $file['StudyGuideID'] ?>">Delete</a></td> 
+                    <th>Display Name</th>
+                    <th>Rename</th>
+                    <th>Delete</th>
                 </tr>
-            <?php } ?>
-        </tbody>
-    </table>
+            </thead>
+            <tbody>
+                <?php foreach ($files as $file) { ?>
+                    <tr>
+                        <td><a target="_blank" class="btn-flat blue-text waves-effect waves-blue no-uppercase" href="<?= $basePath . '/' .$file['FileName'] ?>"><?= $file['DisplayName'] ?></a></td> 
+                        <td><a class="btn waves-effect" 
+                                href="rename-study-guide.php?id=<?= $file['StudyGuideID'] ?>">Rename</a></td> 
+                        <td><a class="btn red white-text waves-effect waves-light" 
+                                href="delete-study-guide.php?id=<?= $file['StudyGuideID'] ?>">Delete</a></td> 
+                    </tr>
+                <?php } ?>
+            </tbody>
+        </table>
+    <?php } ?>
 </div>
 
 <?php include(dirname(__FILE__)."/../footer.php") ?>
